+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 [unreleased]
 
+- Added support for saving files
 - Added support for npy version 2.0
 - Fixed error with empty shapes